place getThingAttributes in more generic location
