+++ src/app/lib/analyticsUtils/article/index.js
@@ -27,21 +27,3 @@ export const getLanguage = articleData =>
 
 export const getPromoHeadline = articleData =>
   pathOr(null, ['promo', 'headlines', 'seoHeadline'], articleData);
-
-export const getThingAttributes = (attribute, articleData) => {
-  const things = pathOr(null, ['metadata', 'tags', 'about'], articleData);
-
-  if (things) {
-    const attributes = [];
-
-    things.forEach(thing => {
-      if (thing[attribute]) {
-        attributes.push(thing[attribute].trim().replace(/\s/g, '+'));
-      }
-    });
-
-    return attributes.join('~') || null;
-  }
-
-  return null;
-};
