+++ src/app/lib/analyticsUtils/article/index.test.js
@@ -3,7 +3,6 @@ const {
   getPageIdentifier,
   getLanguage,
   getPromoHeadline,
-  getThingAttributes,
 } = require('.');
 
 describe('getPageIdentifier', () => {
@@ -129,33 +128,5 @@ describe('getPromoHeadline', () => {
     const promoHeadline = getPromoHeadline(badData);
 
     expect(promoHeadline).toEqual(null);
-  });
-});
-
-describe('getThingAttributes', () => {
-  const data = {
-    metadata: {
-      tags: {
-        about: [{ thingId: 'foo bar' }, { thingId: 'baz' }],
-      },
-    },
-  };
-
-  it('should return thing names in good data', () => {
-    const thingAttributes = getThingAttributes('thingId', data);
-
-    expect(thingAttributes).toEqual('foo+bar~baz');
-  });
-
-  it('should return null if type not found', () => {
-    const thingAttributes = getThingAttributes('fooBar', data);
-
-    expect(thingAttributes).toEqual(null);
-  });
-
-  it('should return null if invalid data', () => {
-    const thingAttributes = getThingAttributes('fooBar', {});
-
-    expect(thingAttributes).toEqual(null);
   });
 });
